+++ NEWS.md
@@ -1,8 +1,14 @@
 NEWS for the htmlTable package
 
+Changes for 2.4.0
+-----------------
+* The `txtRound.data.frame` converted all data to a matrix before rounding all elements. Changed
+  to only convert numeric variables and also added tidy-select syntax. The function now returns
+  the entire input `data.frame` with rounded elements as requested.
+
 Changes for 2.3.0
 -----------------
-* Added `rnames_unique` parameter to tidyHtmlTable so that rows can have non-unique names
+* Added `rnames_unique` parameter to `tidyHtmlTable` so that rows can have non-unique names
 
 Changes for 2.2.1
 -----------------